Per-domain javascript whitelist.
Adds a new `DomainManager` class which handles a new domain.yml file which
currently only supports the `enable-javascript` setting.
`DomainManager.script_toggle()` registers a command to add/remove rules which
are implemented from `WebView.on_url_changed()`. The js policy is applied in
the `WebView` because signals out of tabbed-browser only include the current
tab and this needs to be done in background tabs too.