Adds per-domain cookies white/blacklist.
For whitelist (same origin only) set cookies-accept to never.
For blacklist set cookies-accept to no-3rdparty or all.

Note that removing a domain from whitelist or blacklisting a site will
not remove that domain's cookies from from the cookie jar. But they
won't be served up to the site though.

Implements cookiesForUrl to stop serving cookies up to blocked sites
(for eg if you have changed the policy after already saving a cookie).